display the etag information of the object

+++ integration/s3gateway_test.sh
@@ -15,6 +15,16 @@
 #  limitations under the License.
 
 # environment
+
+os="linux"
+errno=$errno
+if [[ `uname  -a` =~ "Darwin" ]];then
+    os="mac"
+    errno=254
+fi
+echo "os=$os"
+
+
 set -x
 MINT_DATA_DIR=testdata
 MINT_MODE=core
@@ -24,6 +34,9 @@ SECRET_KEY="testUserPassword"
 ENABLE_HTTPS=0
 SERVER_REGION=us-east-1
 ENABLE_VIRTUAL_STYLE=0
+
+
+
 
 # create testdata
 declare -A data_file_map
@@ -44,13 +57,19 @@ data_file_map["datafile-65-MB"]="65M"
 data_file_map["datafile-129-MB"]="129M"
 
 mkdir -p "$MINT_DATA_DIR"
-for filename in "${!data_file_map[@]}"; do
-    echo "creating $MINT_DATA_DIR/$filename"
-    if ! shred -n 1 -s "${data_file_map[$filename]}" - 1>"$MINT_DATA_DIR/$filename" 2>/dev/null; then
-        echo "unable to create data file $MINT_DATA_DIR/$filename"
-        exit 1
-    fi
-done
+
+
+
+# look for empty dir
+if [ ! "$(ls $MINT_DATA_DIR)" ]; then
+    for filename in "${!data_file_map[@]}"; do
+        echo "creating $MINT_DATA_DIR/$filename"
+        if ! shred -n 1 -s "${data_file_map[$filename]}" - 1>"$MINT_DATA_DIR/$filename" 2>/dev/null; then
+            echo "unable to create data file $MINT_DATA_DIR/$filename"
+            exit 1
+        fi
+    done
+fi
 
 # configuration
 aws configure set aws_access_key_id "$ACCESS_KEY"
@@ -81,12 +100,24 @@ fi
 
 
 # test
-HASH_1_KB=$(md5sum "${MINT_DATA_DIR}/datafile-1-kB" | awk '{print $1}')
-HASH_65_MB=$(md5sum "${MINT_DATA_DIR}/datafile-65-MB" | awk '{print $1}')
+function get_md5() {
+    if [ $os == "mac" ]; then
+        md5rt=$(md5 "$1" | awk '{print $4}')
+    else
+        md5rt=$(md5sum "$1" | awk '{print $1}')
+    fi
+}
+
+get_md5 "${MINT_DATA_DIR}/datafile-1-kB"
+HASH_1_KB=$md5rt
+
+get_md5 "${MINT_DATA_DIR}/datafile-65-MB"
+HASH_65_MB=$md5rt
 
 _init() {
     AWS="aws --endpoint-url $1"
 }
+
 
 function get_time() {
     date +%s%N
@@ -216,7 +247,8 @@ function test_upload_object() {
         out=$($function 2>&1)
         rv=$?
         # calculate the md5 hash of downloaded file
-        hash2=$(md5sum /tmp/datafile-1-kB | awk '{print $1}')
+        get_md5 "/tmp/datafile-1-kB"
+        hash2=$md5rt
     fi
 
     # if download succeeds, verify downloaded file
@@ -488,6 +520,16 @@ function test_multipart_upload_0byte() {
     fi
 
     if [ $rv -eq 0 ]; then
+        ret_etag=$(echo "$out" | jq -r .ETag | sed -e 's/^"//' -e 's/"$//')
+        # match etag
+        if [ "$etag" != "$ret_etag" ]; then
+            rv=1
+            out="Etag mismatch for multipart 0 byte object"
+        fi
+        rm -f /tmp/datafile-0-b
+    fi
+
+    if [ $rv -eq 0 ]; then
         function="delete_bucket"
         out=$(delete_bucket "$bucket_name")
         rv=$?
@@ -498,7 +540,7 @@ function test_multipart_upload_0byte() {
     if [ $rv -eq 0 ]; then
         log_success "$(get_duration "$start_time")" "${test_function}"
     else
-        #clean up and log error
+        # clean up and log error
         ${AWS} s3 rb s3://"${bucket_name}" --force > /dev/null 2>&1
         rm -f /tmp/multipart
         log_failure "$(get_duration "$start_time")" "${function}" "${out}"
@@ -653,6 +695,7 @@ function test_max_key_list() {
 
     return $rv
 }
+
 # Copy object tests for server side copy
 # of the object, validates returned md5sum.
 function test_copy_object() {
@@ -679,10 +722,11 @@ function test_copy_object() {
         test_function=${function}
         out=$($function)
         rv=$?
-        if [ $rv -ne 0 ] ; then
+        hash2=$(echo "$out" | jq -r .CopyObjectResult.ETag | sed -e 's/^"//' -e 's/"$//')
+        if [ $rv -eq 0 ] && [ "$HASH_1_KB" != "$hash2" ]; then
             # Verification failed
             rv=1
-            out="test copy-object failed"
+            out="Hash mismatch expected $HASH_1_KB, got $hash2"
         fi
     fi
 
@@ -723,10 +767,18 @@ function test_copy_object_storage_class() {
         test_function=${function}
         out=$($function 2>&1)
         rv=$?
+        # if this functionality is not implemented return right away.
         if [ $rv -ne 0 ]; then
+            if echo "$out" | grep -q "NotImplemented"; then
+                ${AWS} s3 rb s3://"${bucket_name}" --force > /dev/null 2>&1
+                return 0
+            fi
+        fi
+        hash2=$(echo "$out" | jq -r .CopyObjectResult.ETag | sed -e 's/^"//' -e 's/"$//')
+        if [ $rv -eq 0 ] && [ "$HASH_1_KB" != "$hash2" ]; then
             # Verification failed
             rv=1
-            out="test copy object failed"
+            out="Hash mismatch expected $HASH_1_KB, got $hash2"
         fi
     fi
 
@@ -766,12 +818,18 @@ function test_copy_object_storage_class_same() {
         test_function=${function}
         out=$($function 2>&1)
         rv=$?
-
-        # if copy succeeds stat the object
+        # if this functionality is not implemented return right away.
         if [ $rv -ne 0 ]; then
+            if echo "$out" | grep -q "NotImplemented"; then
+                ${AWS} s3 rb s3://"${bucket_name}" --force > /dev/null 2>&1
+                return 0
+            fi
+        fi
+        hash2=$(echo "$out" | jq -r .CopyObjectResult.ETag | sed -e 's/^"//' -e 's/"$//')
+        if [ $rv -eq 0 ] && [ "$HASH_1_KB" != "$hash2" ]; then
             # Verification failed
             rv=1
-            out="test copy object failed"
+            out="Hash mismatch expected $HASH_1_KB, got $hash2"
         fi
     fi
 
@@ -812,7 +870,8 @@ function test_presigned_object() {
         url=$($function)
         rv=$?
         curl -sS -X GET "${url}" > /tmp/datafile-1-kB
-        hash2=$(md5sum /tmp/datafile-1-kB | awk '{print $1}')
+        get_md5 /tmp/datafile-1-kB
+        hash2=$md5rt
         if [ "$HASH_1_KB" == "$hash2" ]; then
             function="delete_bucket"
             out=$(delete_bucket "$bucket_name")
@@ -1134,7 +1193,7 @@ function test_list_objects_error() {
         test_function=${function}
         out=$($function 2>&1)
         rv=$?
-        if [ $rv -ne 255 ]; then
+        if [ $rv -ne $errno ]; then
             rv=1
         else
             rv=0
@@ -1147,7 +1206,7 @@ function test_list_objects_error() {
         test_function=${function}
         out=$($function 2>&1)
         rv=$?
-        if [ $rv -ne 255 ]; then
+        if [ $rv -ne $errno ]; then
             rv=1
         else
             rv=0
@@ -1195,7 +1254,7 @@ function test_put_object_error() {
         test_function=${function}
         out=$($function 2>&1)
         rv=$?
-        if [ $rv -ne 255 ]; then
+        if [ $rv -ne $errno ]; then
             rv=1
         else
             rv=0
@@ -1208,7 +1267,7 @@ function test_put_object_error() {
         test_function=${function}
         out=$($function 2>&1)
         rv=$?
-        if [ $rv -ne 255 ]; then
+        if [ $rv -ne $errno ]; then
             rv=1
         else
             rv=0
@@ -1266,7 +1325,8 @@ function test_serverside_encryption() {
         etag2=$(echo "$out" | jq -r .ETag)
         sse_customer_key2=$(echo "$out" | jq -r .SSECustomerKeyMD5)
         sse_customer_algo2=$(echo "$out" | jq -r .SSECustomerAlgorithm)
-        hash2=$(md5sum /tmp/datafile-1-kB | awk '{print $1}')
+        get_md5 "/tmp/datafile-1-kB"
+        hash2=$md5rt
         # match downloaded object's hash to original
         if [ "$HASH_1_KB" == "$hash2" ]; then
             function="delete_bucket"
@@ -1339,7 +1399,8 @@ function test_serverside_encryption_multipart() {
         etag2=$(echo "$out" | jq -r .ETag)
         sse_customer_key2=$(echo "$out" | jq -r .SSECustomerKeyMD5)
         sse_customer_algo2=$(echo "$out" | jq -r .SSECustomerAlgorithm)
-        hash2=$(md5sum /tmp/datafile-65-MB | awk '{print $1}')
+        get_md5 "${MINT_DATA_DIR}/datafile-65-MB"
+        hash2=$md5rt
         # match downloaded object's hash to original
         if [ "$HASH_65_MB" == "$hash2" ]; then
             function="delete_bucket"
@@ -1450,7 +1511,7 @@ function test_serverside_encryption_multipart_copy() {
         test_function=${function}
         out=$($function)
         rv=$?
-        if [ $rv -ne 255 ]; then
+        if [ $rv -ne $errno ]; then
             rv=1
         else
             rv=0
@@ -1534,7 +1595,7 @@ function test_serverside_encryption_error() {
         rv=$?
     fi
 
-    if [ $rv -ne 255 ]; then
+    if [ $rv -ne $errno ]; then
         rv=1
     else
         rv=0
@@ -1547,7 +1608,7 @@ function test_serverside_encryption_error() {
         rv=$?
     fi
 
-    if [ $rv -ne 255 ]; then
+    if [ $rv -ne $errno ]; then
         rv=1
     else
         rv=0
@@ -1568,7 +1629,7 @@ function test_serverside_encryption_error() {
         out=$($function 2>&1)
         rv=$?
     fi
-    if [ $rv -ne 255 ]; then
+    if [ $rv -ne $errno ]; then
         rv=1
     else
         rv=0
